requirements.txt
MouseInfo==0.1.3
pillow==10.4.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyobjc-core==10.3.1; sys_platform == 'darwin'
pyobjc-framework-Cocoa==10.3.1; sys_platform == 'darwin'
pyobjc-framework-Quartz==10.3.1; sys_platform == 'darwin'
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytweening==1.2.0
rubicon-objc==0.4.9; sys_platform == 'darwin'
python-xlib==0.33; sys_platform == 'linux'
pynput==1.7.6; sys_platform == 'linux'
reportlab==4.0.0
